pyproject.toml
[project]
name = "discovery-github"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "pygithub>=2.4.0",
    "azure-ai-inference>=1.0.0b6",
    "python-dotenv>=1.0.0",
    "openai>=2.3.0",
    "streamlit>=1.50.0",
    "langsmith[openai-agents]>=0.4.34",
    "langfuse>=3.6.2",
    "nest-asyncio>=1.6.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=1.0.1",
    "langchain>=1.0.2",
    "langchain-groq>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-huggingface>=1.0.0",
]
